refactor: rename scope file
