+++ packages/runtime/src/index.ts
@@ -94,7 +94,7 @@ export {
 export {
   BindingContext,
   Scope,
-} from './observation/binding-context';
+} from './observation/scope';
 export {
   CollectionLengthObserver,
   CollectionSizeObserver,
