+++ packages/runtime/src/observation.ts
@@ -1,7 +1,7 @@
 import { DI, IDisposable, IIndexable, IServiceLocator } from '@aurelia/kernel';
 import { isArray } from './utilities-objects';
 
-import type { Scope } from './observation/binding-context';
+import type { Scope } from './observation/scope';
 import type { CollectionLengthObserver, CollectionSizeObserver } from './observation/collection-length-observer';
 import { TaskQueue } from '@aurelia/platform';
